#12467 - Add pathogen tests to environment samples - review changesa and tests fix

+++ sormas-api/src/main/java/de/symeda/sormas/api/environment/environmentsample/EnvironmentSampleIndexDto.java
@@ -97,7 +97,7 @@ public class EnvironmentSampleIndexDto extends PseudonymizableIndexDto implement
 		String otherSampleMaterial,
 		DeletionReason deletionReason,
 		String otherDeletionReason,
-		int numberOfTests,
+		Long numberOfTests,
 		boolean isInJurisdiction) {
 		super(uuid);
 		this.fieldSampleId = fieldSampleId;
